(BOLT-1076) Manually restore the reference to bolt
In order to avoid bolt-vanagon CI failures in https://github.com/puppetlabs/bolt-vanagon/pull/82/commits/fb94b0e750f3090e87fb00e29d4945b4421f2067 a fork of the bolt project was referenced. This allowed the vanagon project to run, but after merging the corresponding PR to Bolt the job to update the reference failed to find the reference BOLT-1076 in the bolt project. This commit manually makes the update that the bolt component pipeline would have made.