(BOLT-1303) Initial work to add a `puppet-bolt` docker container
This uses the same tooling / pattern as the puppet-agent-ubuntu container. It adds documentation for how one might use the bolt container. The focus is on the different ways to pass information to the container and to highlight some of the differences between running bolt locally vs in a container. Analytics has been disabled for the container and a simple spec test to verify running a command in the container works and that the analytics has been disabled.