(GH-2291) Create Bootsnap cache on install on Windows
This commit precompiles the Bootsnap compile cache when Bolt is
installed so that the cache is available on the first Bolt run, and
doesn't lead to that first run taking a long time. Bolt configures the
Bootsnap cache dir to be `$env:ProgramData/Puppetlabs/bolt`, and ensures
the user has permissions to that directory so that the cache can be
updated during Bolt runs after install.