(PE-36810) Stop using dynamic tmp directory for batch script
This commit removes the intermediate script constructed for installation from being written to the %TEMP% directory. Intead it is written to the same directory the script is being run from. This ensures users must be able to write to the directory this script is being run from (for example C:\windows\system32 instead of %TEMP%).