Fix bumping openbolt version
The gemspec seems to have been using the version from Bolt::VERSION
since a long time, no need to update it when releasing a new version
(probably some copy-pasta).

Be more tolerant on what can be found in the version, as we incorrectly
released 5.0.0.rc1 instead of 5.0.0-rc1.  The version string is not
expected to contain `'` char, so it is probably enough to capture what
is in the quotes.