CI: Cleanup linux integration tests
This fixes a regression introduced in
573cdb5454f360306bd5ebd6aa88256724f07a3c /
https://github.com/OpenVoxProject/openbolt/pull/28. We run bolt on a
couple of containers and valid the resultset. A Puppet 6 container was
removed, but the test still expected the result from that. This wasn't
noticed in CI, because the tests were only executed when specific files
changed, and that wasn't triggered in this specific case. We cleaned up
CI and all jobs now run always.