parallel test: change port value to puppet_7_node
Use the port value for puppet_7_node, which is now used in testing
instead of puppet_6_node.

This plan is tightly coupled with the test in
spec/integration/parallel_spec.rb "finishes executing the block then
raises an error when there's an error".

In 573cdb5454f360306bd5ebd6aa88256724f07a3c the puppet 6 node was
removed from testing and the puppet 7 node was used to replace it.
This plan relied on matching the port value of the puppet 6 node to
decide to run a failing task, but the puppet 7 node has a different port
number.